+++ cuip/__init__.py
@@ -1,3 +1,6 @@
+from .weather import *
+from .utils import *
+
 import pkg_resources
 
 try: