+++ cuip/weather/__init__.py
@@ -1,0 +1,1 @@
+#from cuip.cuip.weather import *