+++ CHANGES.rst
@@ -6,3 +6,4 @@ Version 0.1
 ===========
 
 - Initial Commit
+- Added Weather Information scraper